Initial work on Relm4 0.5.0
Added sidebar component
Added factory list
