+++ Cargo.toml
@@ -14,7 +14,7 @@ readme = "README.md"
 
 [dependencies]
 glib = "0.15.11"
-relm4 = { version = "0.4.4", features = ["libadwaita", "macros", "tokio-rt"] }
+relm4 = { git = "https://github.com/Relm4/Relm4", branch = "next", version = "0.5.0-beta.1", features = ["all"] }
 serde = { version = "1.0.136", features = ["derive"] }
 serde_json = "1.0"
 anyhow = "1.0.56"
