+++ src/core/local/tasks.rs
@@ -1,10 +1,10 @@
 use anyhow::Result;
 use diesel::prelude::*;
+use crate::models::queryable::task::QueryableTask;
 
-use crate::models::task::QueryableTask;
+use crate::models::task::Task;
 use crate::schema::tasks::dsl::*;
 use crate::storage::DatabaseConnection;
-use crate::widgets::contents::task_list::Task;
 
 pub fn get_all_tasks() -> Result<Vec<Task>> {
     let connection = DatabaseConnection::establish_connection();
