+++ src/models/queryable/mod.rs
@@ -1,3 +1,2 @@
 pub mod list;
 pub mod task;
-pub mod queryable;
