+++ src/widgets/app.rs
@@ -1,79 +1,42 @@
-use glib::clone;
-use glib::Sender;
-use relm4::{AppUpdate, Components, Model, RelmComponent, send, Widgets};
+use relm4::{Component, ComponentController, ComponentParts, ComponentSender, Controller, SimpleComponent};
 use relm4::gtk::prelude::GtkWindowExt;
 
 use crate::{adw, adw::prelude::AdwApplicationWindowExt, gtk, gtk::gio, gtk::prelude::{BoxExt, ButtonExt, OrientableExt, WidgetExt}};
-use crate::widgets::contents::content::{ContentModel, ContentMsg};
-use crate::widgets::panel::sidebar::{SidebarModel, SidebarMsg};
-use crate::widgets::panel::theme_selector::ThemeSelector;
-use crate::widgets::popovers::new_list::NewListModel;
+use crate::widgets::sidebar::{SidebarInput, SidebarModel, SidebarOutput};
 
 pub struct AppModel {
-    message: Option<AppMsg>,
+    message: Option<Input>,
+    sidebar: Controller<SidebarModel>,
 }
 
 impl AppModel {
-    pub fn new() -> Self {
+    pub(crate) fn new() -> Self {
         Self {
-            message: None
+            message: None,
+            sidebar: SidebarModel::builder()
+                .launch(None)
+                .connect_receiver(move |sender, message| match message {
+                    SidebarOutput::ListSelected(index) => println!("Current index: {}", index)
+                }),
         }
     }
 }
 
-pub enum AppMsg {
-    UpdateContent(ContentMsg),
-    UpdateSidebar(SidebarMsg),
+pub enum Input {
+    AddList(String),
     Folded,
     Unfolded,
     Forward,
     Back,
 }
 
-impl Model for AppModel {
-    type Msg = AppMsg;
+#[relm4::component(pub)]
+impl SimpleComponent for AppModel {
+    type Input = Input;
+    type Output = ();
+    type InitParams = Option<Input>;
     type Widgets = AppWidgets;
-    type Components = AppComponents;
-}
 
-impl AppUpdate for AppModel {
-    fn update(
-        &mut self,
-        msg: Self::Msg,
-        components: &Self::Components,
-        _sender: Sender<Self::Msg>,
-    ) -> bool {
-        match msg {
-            AppMsg::UpdateContent(msg) => components.content.send(msg).unwrap(),
-            AppMsg::UpdateSidebar(msg) => components.sidebar.send(msg).unwrap(),
-            _ => self.message = Some(msg),
-        }
-        true
-    }
-}
-
-pub struct AppComponents {
-    sidebar: RelmComponent<SidebarModel, AppModel>,
-    content: RelmComponent<ContentModel, AppModel>,
-    new_list: RelmComponent<NewListModel, AppModel>,
-    theme_selector: RelmComponent<ThemeSelector, AppModel>,
-}
-
-impl Components<AppModel> for AppComponents {
-    fn init_components(parent_model: &AppModel, parent_sender: Sender<AppMsg>) -> Self {
-        AppComponents {
-            sidebar: RelmComponent::new(parent_model, parent_sender.clone()),
-            content: RelmComponent::new(parent_model, parent_sender.clone()),
-            new_list: RelmComponent::new(parent_model, parent_sender.clone()),
-            theme_selector: RelmComponent::new(parent_model, parent_sender),
-        }
-    }
-
-    fn connect_parent(&mut self, _parent_widgets: &AppWidgets) {}
-}
-
-#[relm4::widget(pub)]
-impl Widgets<AppModel, ()> for AppWidgets {
     view! {
         window = adw::ApplicationWindow {
             set_default_width: 800,
@@ -118,15 +81,16 @@ impl Widgets<AppModel, ()> for AppWidgets {
                                         },
                                     },
                                 },
-                                pack_start: new_list_button = &gtk::MenuButton {
+                                pack_start: new_list_button = &gtk::Button {
                                     set_icon_name: "value-increase-symbolic",
                                     add_css_class: "raised",
                                     set_has_frame: true,
-                                    set_direction: gtk::ArrowType::None,
-                                    set_popover: Some(components.new_list.root_widget()),
+                                    connect_clicked(sender) => move |_| {
+                                        sender.input(Input::AddList("Test".to_string()));
+                                    }
                                 },
                             },
-                            append: components.sidebar.root_widget()
+                            append: model.sidebar.widget()
                         },
                         append: &gtk::Separator::default(),
                         append: content = &gtk::Box {
@@ -140,20 +104,20 @@ impl Widgets<AppModel, ()> for AppWidgets {
                                     set_icon_name: "go-previous-symbolic",
                                     set_visible: false,
                                     connect_clicked(sender) => move |_| {
-                                        send!(sender, AppMsg::Back);
+                                        // sender.send(Input::Back);
                                     }
                                 },
                             },
-                            append: components.content.root_widget()
+                            // append: components.content.root_widget()
                         },
-                        connect_folded_notify: clone!(@weak content, @strong sender => move |leaflet| {
-                            if leaflet.is_folded() {
-                                leaflet.set_visible_child(&content);
-                                send!(sender, AppMsg::Folded);
-                            } else {
-                                send!(sender, AppMsg::Unfolded);
-                            }
-                        })
+                        // connect_folded_notify: clone!(@weak content, @strong sender => move |leaflet| {
+                        //     if leaflet.is_folded() {
+                        //         leaflet.set_visible_child(&content);
+                        //         send!(sender, Input::Folded);
+                        //     } else {
+                        //         send!(sender, Input::Unfolded);
+                        //     }
+                        // })
                     }
                 }
             },
@@ -161,23 +125,37 @@ impl Widgets<AppModel, ()> for AppWidgets {
     }
 
     fn post_view() {
+        println!("Test");
         if let Some(msg) = &model.message {
-            match msg {
-                AppMsg::Folded => {
-                    self.leaflet.set_visible_child(&self.content);
-                    self.go_back_button.set_visible(true);
-                    sidebar_header.set_show_start_title_buttons(true);
-                    sidebar_header.set_show_end_title_buttons(true);
-                }
-                AppMsg::Unfolded => {
-                    self.go_back_button.set_visible(false);
-                    sidebar_header.set_show_start_title_buttons(false);
-                    sidebar_header.set_show_end_title_buttons(false);
-                }
-                AppMsg::Forward => self.leaflet.set_visible_child(&self.content),
-                AppMsg::Back => self.leaflet.set_visible_child(&self.sidebar),
-                _ => {}
-            }
+            // match msg {
+            //     AppMsg::Folded => {
+            //         self.leaflet.set_visible_child(&self.content);
+            //         self.go_back_button.set_visible(true);
+            //         sidebar_header.set_show_start_title_buttons(true);
+            //         sidebar_header.set_show_end_title_buttons(true);
+            //     }
+            //     AppMsg::Unfolded => {
+            //         self.go_back_button.set_visible(false);
+            //         sidebar_header.set_show_start_title_buttons(false);
+            //         sidebar_header.set_show_end_title_buttons(false);
+            //     }
+            //     AppMsg::Forward => self.leaflet.set_visible_child(&self.content),
+            //     AppMsg::Back => self.leaflet.set_visible_child(&self.sidebar),
+            //     _ => {}
+            // }
+        }
+    }
+
+    fn init(params: Self::InitParams, root: &Self::Root, sender: &ComponentSender<Self>) -> ComponentParts<Self> {
+        let model = AppModel::new();
+        let widgets = view_output!();
+        ComponentParts { model, widgets }
+    }
+
+    fn update(&mut self, message: Self::Input, sender: &ComponentSender<Self>) {
+        match message {
+            Input::AddList(title) => self.sidebar.sender().send(SidebarInput::AddList(title)),
+            _ => self.message = Some(message),
         }
     }
 }
