--- src/widgets/contents/mod.rs
@@ -1,3 +1,0 @@
-pub mod content;
-pub mod details;
-pub mod task_list;
