+++ src/widgets/mod.rs
@@ -1,4 +1,3 @@
 pub mod app;
-pub mod contents;
-pub mod panel;
-pub mod popovers;
+mod sidebar;
+mod list;
