--- src/widgets/panel/mod.rs
@@ -1,3 +1,0 @@
-pub mod list;
-pub mod sidebar;
-pub mod theme_selector;
