--- src/widgets/popovers/main_menu.rs
@@ -1,1 +1,0 @@
-
