--- src/widgets/popovers/mod.rs
@@ -1,4 +1,0 @@
-pub mod edit_list;
-pub mod edit_task;
-pub mod main_menu;
-pub mod new_list;
